py_projects/ai_smm_agent_1/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.7
requests==2.31.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
aiofiles==23.2.1
pillow==10.1.0
python-telegram-bot==20.7
tweepy==4.14.0
facebook-sdk==3.1.0
instagram-private-api==1.6.0
schedule==1.2.0
asyncio==3.4.3
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
jinja2==3.1.2
markdown==3.5.1
beautifulsoup4==4.12.2
lxml==4.9.3
nltk==3.8.1
textblob==0.17.1
aiohttp==3.9.1
